+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - 2020-12-28
 ### Added
 - Added pre-generated bindings for `i686-unknown-linux` targets.
 - Added pre-generated bindings for the `i686-pc-windows-msvc` target.
@@ -21,5 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed `build.rs` script to handle cross compilation.
 
-[Unreleased]: https://github.com/newAM/libftd2xx-ffi-rs/compare/0.4.1...HEAD
+[Unreleased]: https://github.com/newAM/libftd2xx-ffi-rs/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/newAM/libftd2xx-ffi-rs/compare/0.5.0...0.4.1
 [0.4.1]: https://github.com/newAM/libftd2xx-ffi-rs/compare/0.4.0...0.4.1